+++ src/fs/delete.js
@@ -1,7 +1,7 @@
 import { unlink } from "fs/promises";
 import { join } from "path";
 
-import { exists, pathToDir } from "./shared.js";
+import { exists, pathToDir } from "../shared.js";
 
 export const remove = async () => {
   const __dirname = pathToDir(import.meta.url);
