+++ src/streams/read.js
@@ -1,9 +1,10 @@
 import { createReadStream } from "fs";
-import { dirname, join } from "path";
-import { fileURLToPath } from "url";
+import { join } from "path";
+
+import { pathToDir } from "../shared.js";
 
 export const read = async () => {
-  const __dirname = dirname(fileURLToPath(import.meta.url));
+  const __dirname = pathToDir(import.meta.url);
   const __filename = "files/fileToRead.txt";
   const readStream = createReadStream(join(__dirname, __filename));
