+++ src/streams/write.js
@@ -1,9 +1,10 @@
 import { createWriteStream } from "fs";
-import { dirname, join } from "path";
-import { fileURLToPath } from "url";
+import { join } from "path";
+
+import { pathToDir } from "../shared.js";
 
 export const write = async () => {
-  const __dirname = dirname(fileURLToPath(import.meta.url));
+  const __dirname = pathToDir(import.meta.url);
   const __filename = "files/fileToWrite.txt";
   const writeStream = createWriteStream(join(__dirname, __filename));
